feat(models/controllers): added noctificaton sys to post comment

+++ backend/controllers/follow.js
@@ -1,9 +1,31 @@
 const Follow = require('./../models/follow')
+const { Op } = require('sequelize')
+
+exports.checkFollow = (req, res)=>{
+    const from = req.body.from
+    const to = req.body.to
+    console.log(from, to)
+    if(req.session.userId !== from){
+        res.status(401).json({message:"vérification non autorisé !"})
+    }
+    else{
+        Follow.findOne({where:{[Op.and]:[{from: from},{to: to}]}})
+        .then(result => {
+            console.log(result)
+            if (result === null){
+                res.status(204).json({result: "non suivi"})
+            }
+            else{
+                res.status(200).json({result: "suivi"})
+            }
+        })
+        .catch(error => console.log(error) )
+    }
+}
 
 exports.followOne = (req,res)=>{
     const from = req.body.from
     const to = req.body.to
-    console.log(req.body)
     if(to === from){
         res.status(401).json({message:'non autorisé'})
     }
@@ -18,9 +40,14 @@ exports.followOne = (req,res)=>{
 exports.deleteFollow = (req,res)=>{
     const from = req.body.from
     const to = req.body.to
+    if(from == undefined || to == undefined){
+        return res.status(400).json({message:'mauvais corps de requete'})
+    }
     Follow.destroy({
         where:{
             [Op.and]:[{from},{to}]
         }
     })
+    .then(()=> res.status(200).json({message:'vous ne le suivez plus !'}))
+    .catch(error => res.status(400).json({error}))
 }
